+++ ppsp.py
@@ -1,20 +1,13 @@
 import os, sys, queue, re, subprocess, threading
 
 
-class PPSP:
-    class Exceptions:
-        class PPSPException(Exception):
-            pass
-        
-        class PPSPExitException(PPSPException):
-            pass
-
+class _PPSP:
     class _Status:
         running = False
         stopping = False
         stopped = True
 
-    def __init__(self, shell_command: str, verbose: bool = False, exit_condition: str = None):
+    def __init__(self, shell_command: str, exit_condition: str = None):
         # Subprocess management
         self._shell_command = shell_command
 
@@ -22,13 +15,10 @@ class PPSP:
             self._exit_condition = re.compile(exit_condition) \
                 if exit_condition is not None else None
         except re.error as e:
-            if verbose:
-                print('Regular expression did not compile. Skipping...')
+            pass
             
             self._exit_condition = None
         
-        self._exit_condition_met = False
-
         # Subprocess
         self._subprocess = None
 
@@ -38,15 +28,28 @@ class PPSP:
         self._writer_tr = threading.Thread(target=self.__writer)
 
         self._status = self._Status()
-        self._queue = queue.Queue()
+        self._stdin_queue = queue.Queue()
+        self._stdout_queue = queue.Queue()
+
+    @property
+    def shell_command(self):
+        return self._shell_command
 
     @property
     def subprocess(self):
         return self._subprocess
 
+    @property
+    def status(self):
+        return self._status
+
+    @property
+    def stdout_queue(self):
+        return self._stdout_queue
+
     def process_command(self, cmd: str) -> None:
         if cmd:
-            self._queue.put(cmd)
+            self._stdin_queue.put(cmd)
 
     def start(self) -> None:
         self.__start_process()
@@ -67,17 +70,11 @@ class PPSP:
         self._status.stopping = False
         self._status.stopped = True
 
-        if self._exit_condition_met:
-            raise PPSP.Exceptions.PPSPExitException
-
-    def is_running(self) -> bool:
-        return self._status.running
-    
-    def is_stopping(self) -> bool:
-        return self._status.stopping
-
-    def is_stopped(self) -> bool:
-        return self._status.stopped
+        if self._subprocess is not None:
+            try:
+                self._subprocess.kill()
+            except (OSError, ProcessLookupError):
+                pass
 
     def __start_process(self) -> None:
         self._subprocess = subprocess.Popen(self._shell_command.split(), 
@@ -113,10 +110,10 @@ class PPSP:
                         break
                     line = linen.decode(sys.getdefaultencoding())
                     if line:
-                        print(line, end='')
+                        self._stdout_queue.put(line.strip())
+
                         if self._exit_condition:
                             if re.match(self._exit_condition, line):
-                                self._exit_condition_met = True
                                 self._status.stopping = True
 
     def __writer(self) -> None:
@@ -127,60 +124,24 @@ class PPSP:
             if self._status.stopping:
                 break
 
-            if not self._queue.empty():
-                self.__send_input(self._queue.get())
+            if not self._stdin_queue.empty():
+                self.__send_input(self._stdin_queue.get())
 
 
-def ppsp(shell_command: str, verbose: bool = False, exit_condition: str = None) -> None:
-    if verbose:
-        print('PPSP starting... Press CTRL+C to quit')
+def __ppsp(shell_command: str, return_val: dict = {}, exit_condition: str = None) -> None:
+    ppsp_instance = _PPSP(shell_command, exit_condition)
+    return_val['val'] = ppsp_instance
 
-    ppsp_instance = None
-    try:
-        ppsp_instance = PPSP(shell_command, verbose, exit_condition)
-        ppsp_instance.start()
-        ppsp_instance.stop()
-    except KeyboardInterrupt:
-        if verbose:
-            print('Keyboard Interrupt. Exiting PPSP gracefully...')
 
-        if ppsp_instance is not None:
-            if ppsp_instance.subprocess is not None:
-                try:
-                    ppsp_instance.subprocess.kill()
-                except (OSError, ProcessLookupError):
-                    pass
-        sys.exit(0)
-    except PPSP.Exceptions.PPSPExitException:
-        if verbose:
-            print('Exit condition met. Exiting PPSP gracefully...')
+def ppsp(shell_command: str, exit_condition: str = None) -> _PPSP:
+    return_val = {
+        'val': None
+    }
+    ppsp_th = threading.Thread(target=__ppsp, args=(shell_command, return_val, exit_condition),
+        daemon=True)
+    ppsp_th.start()
 
-        if ppsp_instance is not None:
-            if ppsp_instance.subprocess is not None:
-                try:
-                    ppsp_instance.subprocess.kill()
-                except (OSError, ProcessLookupError):
-                    pass
-        sys.exit(0)
-    except Exception:
-        if verbose:
-            print('Unhandled exception occured. Exiting PPSP gracefully...')
-        
-        if ppsp_instance is not None:
-            if ppsp_instance.subprocess is not None:
-                try:
-                    ppsp_instance.subprocess.kill()
-                except (OSError, ProcessLookupError):
-                    pass
-        sys.exit(0)
-    else:
-        if verbose:
-            print('Finished normally. Exiting PPSP gracefully...')
-        
-        if ppsp_instance is not None:
-            if ppsp_instance.subprocess is not None:
-                try:
-                    ppsp_instance.subprocess.kill()
-                except (OSError, ProcessLookupError):
-                    pass
-        sys.exit(0)
+    while return_val['val'] is None:
+        continue
+
+    return return_val['val']